+++ packages/cli/src/ui/utils/textUtils.ts
@@ -244,6 +244,45 @@ export function sliceTextByVisualHeight(
 }
 
 /**
+ * Wrap text into the visual rows it occupies at `width` columns, accounting
+ * for both explicit newlines and code-point-width-aware soft wrapping. Unlike
+ * `sliceTextByVisualHeight` (which keeps only a head/tail window), this returns
+ * every visual row, so callers that scroll an arbitrary offset (e.g. the
+ * ThinkingViewer) can slice the rows the user actually sees.
+ */
+export function wrapToVisualLines(text: string, width: number): string[] {
+  if (width <= 0) {
+    return [''];
+  }
+  const visualLines: string[] = [];
+  for (const logicalLine of text.split('\n')) {
+    if (logicalLine === '') {
+      visualLines.push('');
+      continue;
+    }
+    let currentLine = '';
+    let currentWidth = 0;
+    for (const char of logicalLine) {
+      const charWidth = getCachedStringWidth(char);
+      if (currentWidth + charWidth > width && currentWidth > 0) {
+        visualLines.push(currentLine);
+        currentLine = '';
+        currentWidth = 0;
+      }
+      currentLine += char;
+      currentWidth += charWidth;
+    }
+    if (currentLine) {
+      visualLines.push(currentLine);
+    }
+  }
+  if (visualLines.length === 0) {
+    visualLines.push('');
+  }
+  return visualLines;
+}
+
+/**
  * Clear the string width cache
  */
 export const clearStringWidthCache = (): void => {
